Slider markup and styling on Property Page
I marked up a slider in the Values Appreciation section of the Property
page. Also styled it. Next, it has to be made draggable with adjustable
values. And this has to be duplicated to the next two slider areas.
After this, the Property page on its own is done and I will draw out the
subpages for the different sections.