Signup Page beginnings creation and typography fix
Typography: Chrome's new update prevents the loading of Segoe UI Light
fonts, this makes it necessary to ensure that all the fallback fonts are
working properly. We have Proxima Nova from Adobe TypeKit As the main
fallback font. It is set to work on http://localhost/ as well as
Purchis.com so you can test it out locally.

Sign Up: Created the beginnings of the Sign up form process. There is
still the main form and agreement to implement.